[Refactor/#7] CreateForm의 View를 분리
